# server/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.2"
  title: 18F
# during dev, should point to your local machine
#host: localhost:10010
# basePath prefixes all resource paths 
basePath: /
# 
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /search/splash:
    get:
      x-swagger-router-controller: fda_api
      operationId: aggregateSplashSearch
      description: Aggregated search for the splash page
      parameters:
        - name: query
          in: query
          description: Value to search on in hardcoded FDA endpoint for the sites splash page
          required: true
          type: string
        - name: drugFields
          in: query
          description: The fields one wishes to search in the drug results. Comma seperated list.
          required: false
          type: string
        - name: foodFields
          in: query
          description: The fields one wishes to search in the food results. Comma seperated list.
          required: false
          type: string
        - name: deviceFields
          in: query
          description: The fields one wishes to search in the device results. Comma seperated list.
          required: false
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/Generic"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /drug/event/search:
    get:
      x-swagger-router-controller: fda_api
      operationId: drugEventSearch
      description: Search field(s)
      parameters:
        - name: query
          in: query
          description: Field and Value to search on separated by a colon ( field:value ). query string should be in quotes. Format example= primarysourcecountry:us
          required: true
          type: string
          collectionFormat: multi
        - name: skip
          description: Number of items to skip and start results from
          in: query
          required: false
          type: integer
        - name: limit
          in: query
          description: Limit the number of results
          required: false
          type: integer
      produces:
        - application/json
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/Generic"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /drug/event/rangecount:
    get:
      x-swagger-router-controller: fda_api
      operationId: drugEventRangeCount
      description: Returns test fetch of a drugs query count
      parameters:
        - name: start
          in: query
          description: Start day in format YYYYMMDD (ex 20040101)
          required: true
          type: string
        - name: end
          in: query
          description: End day in format YYYYMMDD (ex 20150101)
          required: true
          type: string
        - name: field
          in: query
          description: path of field to count (ex patient.patientsex)
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/Generic"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"

  /device/event/search:
    get:
      x-swagger-router-controller: fda_api
      operationId: deviceEventSearch
      description: Search field(s)
      parameters:
        - name: query
          in: query
          description: Field and Value to search on separated by a colon ( field:value ). query string should be in quotes. Format example= primarysourcecountry:us
          required: true
          type: string
          collectionFormat: multi
        - name: skip
          description: Number of items to skip and start results from
          in: query
          required: false
          type: integer
        - name: limit
          in: query
          description: Limit the number of results
          required: false
          type: integer
      produces:
        - application/json
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/Generic"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /device/event/rangecount:
    get:
      x-swagger-router-controller: fda_api
      operationId: deviceEventRangeCount
      description: Returns test fetch of a drugs query count
      parameters:
        - name: start
          in: query
          description: Start day in format YYYYMMDD (ex 20040101)
          required: true
          type: string
        - name: end
          in: query
          description: End day in format YYYYMMDD (ex 20150101)
          required: true
          type: string
        - name: field
          in: query
          description: path of field to count (ex patient.patientsex)
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/Generic"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"

  /food/event/search:
    get:
      x-swagger-router-controller: fda_api
      operationId: foodEventSearch
      description: Search field(s)
      parameters:
        - name: query
          in: query
          description: Field and Value to search on separated by a colon ( field:value ). query string should be in quotes. Format example= primarysourcecountry:us
          required: true
          type: string
          collectionFormat: multi
        - name: skip
          description: Number of items to skip and start results from
          in: query
          required: false
          type: integer
        - name: limit
          in: query
          description: Limit the number of results
          required: false
          type: integer
      produces:
        - application/json
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/Generic"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /food/event/rangecount:
      get:
        x-swagger-router-controller: fda_api
        operationId: foodEventRangeCount
        description: Returns test fetch of a drugs query count
        parameters:
          - name: start
            in: query
            description: Start day in format YYYYMMDD (ex 20040101)
            required: true
            type: string
          - name: end
            in: query
            description: End day in format YYYYMMDD (ex 20150101)
            required: true
            type: string
          - name: field
            in: query
            description: path of field to count (ex patient.patientsex)
            required: true
            type: string
        responses:
          "200":
            description: Success
            schema:
              # a pointer to a definition
              $ref: "#/definitions/Generic"
          # responses may fall through to errors
          default:
            description: Error
            schema:
              $ref: "#/definitions/ErrorResponse"

# complex objects have schema definitions
definitions:
  Generic:
    properties:
      message:
        type: string
  ErrorResponse:
      required:
        - code
        - message
      properties:
        code:
          type: integer
        message:
          type: string